crud - I broke the API into smaller files which are composed in SeedsApi.
